Refactor code and fix urls routes - Refactor routers
