+++ frontend/src/components/SharedUsersHeader/SharedUsersHeader.tsx
@@ -1,12 +1,16 @@
 import React from 'react'
-import { Typography, Tooltip, IconButton } from '@material-ui/core'
+import { Typography, IconButton } from '@material-ui/core'
 import { Breadcrumbs } from '../Breadcrumbs'
 import { Icon } from '../Icon'
 import { Title } from '../Title'
-type Props = {
-    onClick?: ()=> void
-  }
-export const SharedUsersHeader:  React.FC<Props> = ({ onClick }) => {
+import { useLocation, useHistory } from 'react-router-dom'
+
+export const SharedUsersHeader = () => {
+
+  const location = useLocation()
+  const history = useHistory()
+
+  const onClick = () => history.push(`${location.pathname}/share`)
 
   return (
     <div>
@@ -15,13 +19,11 @@ export const SharedUsersHeader:  React.FC<Props> = ({ onClick }) => {
       <Typography variant="h1">
         <Icon name="user-friends" size="lg" />
         <Title>Shared users</Title>
-        <Tooltip title="Add User">
-          <div>
+        <div>
           <IconButton onClick={onClick}>
             <Icon name="user-plus" size="md" type="light" />
           </IconButton>
-          </div>
-        </Tooltip>
+        </div>
       </Typography>
       </>
     </div>
