+++ frontend/src/pages/UsersPageService/UsersPageService.tsx
@@ -4,36 +4,27 @@ import { findService } from '../../models/devices'
 import { useSelector } from 'react-redux'
 import { ApplicationState } from '../../store'
 import { Container } from '../../components/Container'
-import { useHistory } from 'react-router-dom'
 import analytics from '../../helpers/Analytics'
 import { SharedUsersList } from '../../components/SharedUsersList'
 import { SharedUsersHeader } from '../../components/SharedUsersHeader'
 
 export const UsersPageService: React.FC = () => {
   const { serviceID = '' } = useParams()
-  const { deviceID = '' } = useParams()
-  const { device, service } = useSelector((state: ApplicationState) => ({
-    device: state.devices.find((d: IDevice) => d.id === deviceID),
-    service: findService(state.devices.all, serviceID)[0]
-  }))
+  const [service, device] = useSelector((state: ApplicationState) => findService(state.devices.all, serviceID))
   const shared =  service?.access.length
-
-  const history = useHistory()
 
   useEffect(() => {
     analytics.page('UsersPageService')
   }, [])
 
-  const onClick = () => history.push(`/devices/${deviceID}/share`)
-
   return (
     <Container
       scrollbars
       header={
-        <SharedUsersHeader onClick={onClick} />
+        <SharedUsersHeader />
       }
     >
-     {(service && shared) && <SharedUsersList device={device} service={service} />}
+     {(service && device && shared) && <SharedUsersList device={device} service={service} />}
     </Container>
   )
 }
